Add files via upload
An AUC (Area Under the Curve) value of 0.51 indicates that the performance of the model is slightly better than random chance. The AUC ranges from 0 to 1, where a value of 0.5 indicates a random classifier and a value of 1 indicates a perfect classifier.

In the case of an AUC value of 0.51, it suggests that the model's ability to distinguish between positive and negative samples is only slightly better than randomly guessing. This means that the model's predictive power is limited, and its performance may not be sufficient for accurate predictions or classifications.

It's important to note that an AUC value of 0.51 does not necessarily mean that the model is entirely ineffective. It may still provide some useful insights or be a starting point for further improvements. However, it indicates that there is room for enhancement and the model's performance should be evaluated further to determine its usefulness in practical applications.